修改性能对比的逻辑
